Fix first wildchar complete only common prefix
Prepending "longest:full" to 'wildmode' before the default value of "full"
means that the string is completed only up to the longest common prefix
when using 'wildchar' for the first time. After that it works like
normal.

With "longest:full" instead of "longest", 'wildmenu' is started even for
the first use.