Set correct dir for files Emacs is told to visit
Commit a135880 (Replace Projectile with frame-local CWD, 2021-02-22)
introduced a mechanism for frame-local default directories. This was
done by setting a CWD frame property in a server-after-make-frame
hook, and then using that in a find-file hook as the default-directory
value for the new buffer. However when Emacs is told to visit some
files like

    emacsclient files ...

the find-file hook gets run for each file before the creation of the
frame is finished. Therefore the inherited CWD would incorrectly be
used instead.

This commit fixes the issue by, in the server-after-make-frame hook,
resetting default-directory to the correct value for the affected
buffers.