Resolve breaking change in tmux 3.2
In order to make the status bar transparent, the status-bg variable
was set to the default background color. However, with tmux 3.2 this
no longer has the intended effect, as mentioned in the release post:

> status-fg and status-bg are now applied before status-style rather
> than instead of; this means that if either is set to default the
> status-style colour will be used. To fix, replace status-bg default
> with status-style bg=default.

This commit fixes the issue by applying the suggestion.

See: https://github.com/tmux/tmux/issues/2645