Make xterm bracketed paste text not get copied
The function `current-kill` is used to get the contents of the OS
clipboard. It, however, adds the result on top of the Emacs kill ring.
This is unnecessary because the desktop environment clipboard manager
already maintains the list of copied items.

Since `kill-new` does not destructively modify the tail of the kill ring
we can advise `current-kill` to save and restore a shallow copy. Another
solution is to override `current-kill` only when it otherwise would end
up pushing to the kill ring:

    (advice-add 'current-kill :around
                (lambda (orig-fun n &rest args)
                  (let ((interprogram-paste (and (= n 0) interprogram-paste-function
                                                 (funcall interprogram-paste-function))))
                    (or (car-safe interprogram-paste) interprogram-paste
                        (apply orig-fun n args)))))

but this is more finicky.