Use motion as initial state for read-only modes
The keys "ZZ"/"ZQ" get reserved in evil-motion-state-map for their
normal mode commands. This means they would have to be overwritten
individually in both the visual/operator states for actual motions to be
bound to them. It should be noted that there is precedent for this in
Evil; `evil-scroll-line-to-center` is bound to "zz" in motion state.

evil-collection instead uses normal state for this purpose and remaps
insertion commands to be ignored. While less obtrusive, this requires
rebinding all the default keys for each read-only mode, which is too
tedious.