Move only vertically on gj/gk despite tracking EOL
With the gj/gk motions the intention is to move some number of display
lines downward or upward. If there is a character visually directly
above/below the cursor, then you do not expect the cursor to also move
horizontally, even if the cursor was tracking the end of the line.