Ditch special keymap for editing rebase sequences
The benefits do not outweigh the effort required to learn to use it,
plus I enjoy using the vi "d" operator to reorder multiple commands.

The suggestion to try to disable git-rebase-mode completely with

    (setq auto-mode-alist (rassq-delete-all 'git-rebase-mode auto-mode-alist))

does not fully solve the problem, since magit-rebase-interactive still
needs with-editor-mode, and git-rebase sometimes sources itself again,
annoyingly readding it to auto-mode-alist.