Use Normal as initial state for read-only modes
This is a reversal on the stance taken in commit
9c8212b90f47eff645854e6b4d6074f67751e32f. The problem is that
mode-specific keybindings for Motion state, which end up in Evil
Auxiliary keymaps, shadow global Evil motions when in Operator-Pending
state. This is because Motion state keymaps are enabled in
Operator-Pending state, to provide motions in addition to the text
objects in Operator-Pending state keymaps.

One possible future enhancement is to create a separate special state,
one that is not associated with Operator-Pending state. Could also make
it easier to copy over default mode map keybindings that do not conflict
with global motions.